chat

+++ resources/views/lk/org.blade.php
@@ -13,7 +13,9 @@
                     <dt>Phone</dt>
                     <dd>{{$org->phone_number}}</dd>
                 </dl>
-                <a href="/org-orders/{{$org->id}}" class="btn btn-success">Заказы</a>
+                @if(Auth::user()->id === $org->user_id)
+                    <a href="/org-orders/{{$org->id}}" class="btn btn-success">Заказы</a>
+                @endif
                 <ul class="list-group my-2">
                     <li class="list-group-item">{{$org::TYPES[$org->type]}}</li>
                     <li class="list-group-item">Адрес: {{$org->jur_address}}</li>
@@ -22,8 +24,9 @@
                     <li class="list-group-item">ОГРН: {{$org->ogrn}}</li>
                     <li class="list-group-item">р/с: {{$org->payment_account}}</li>
                 </ul>
-                <a href="#!" class="btn-link" data-toggle="modal" data-target=".modal-data">Редактировать данные организации</a>
-
+                @if(Auth::user()->id === $org->user_id)
+                    <a href="#!" class="btn-link" data-toggle="modal" data-target=".modal-data">Редактировать данные организации</a>
+                @endif
                 <h4 class="card-title text-center">Цены</h4>
                 <table class="table">
                     <thead class="thead-light">
@@ -43,11 +46,13 @@
                 </table>
                 <h4 class="card-title text-center">О нас</h4>
                 <div class="bg-secondary text-white p-3">{{$org->description}}</div>
-                <div class="card-body d-flex justify-content-between">
-                    <a href="#!" class="btn-link" data-toggle="modal" data-target=".modal-prices">Редактировать цены</a>
-                    <a href="/delete-org/{{$org->id}}" class="btn-link text-danger">Удалить организацию</a>
-                </div>
 
+                @if(Auth::user()->id === $org->user_id)
+                    <div class="card-body d-flex justify-content-between">
+                        <a href="#!" class="btn-link" data-toggle="modal" data-target=".modal-prices">Редактировать цены</a>
+                        <a href="/delete-org/{{$org->id}}" class="btn-link text-danger">Удалить организацию</a>
+                    </div>
+                @endif
             </div>
 
         </div>
